Let main window take care of message boxes and prompts

+++ UpdateChecker.cs
@@ -41,49 +41,32 @@ namespace mgmtapplauncher2
 				Version current = System.Reflection.Assembly.GetExecutingAssembly().GetName().Version;
 				Version latest = new Version(m_LatestVersion);
 
-				if (current.CompareTo(latest) < 0)
+				if (current.CompareTo(latest) < 0 && m_Parent.AskToUpdate(latest.ToString()))
 				{
-
-					MessageBoxResult mbr = MessageBox.Show(
-						String.Format(Strings.MessageUpdateAvailable, latest, App.GetName()),
-						App.GetName(),
-						MessageBoxButton.YesNo,
-						MessageBoxImage.Question
-					);
-
-					if (mbr == MessageBoxResult.Yes)
+					try
 					{
-						try
-						{
 
-							string tempinstaller = Path.GetTempFileName();
-							string installer = Path.GetDirectoryName(tempinstaller) + "\\mgmtapplauncher2.msi";
+						string tempinstaller = Path.GetTempFileName();
+						string installer = Path.GetDirectoryName(tempinstaller) + "\\mgmtapplauncher2.msi";
 
-							new WebClient().DownloadFile("http://updates.kempniu.pl/mgmtapplauncher2/mgmtapplauncher2.msi", tempinstaller);
+						new WebClient().DownloadFile("http://updates.kempniu.pl/mgmtapplauncher2/mgmtapplauncher2.msi", tempinstaller);
 
-							if (File.Exists(installer))
-								File.Delete(installer);
-							File.Move(tempinstaller, installer);
+						if (File.Exists(installer))
+							File.Delete(installer);
+						File.Move(tempinstaller, installer);
 
-							Process updater = new Process();
-							updater.StartInfo.FileName = "msiexec.exe";
-							updater.StartInfo.Arguments = "/i \"" + installer + "\" REINSTALL=ALL REINSTALLMODE=vomus";
-							updater.Start();
+						Process updater = new Process();
+						updater.StartInfo.FileName = "msiexec.exe";
+						updater.StartInfo.Arguments = "/i \"" + installer + "\" REINSTALL=ALL REINSTALLMODE=vomus";
+						updater.Start();
 
-							m_Parent.Close();
+						m_Parent.Close();
 
-						}
-						catch (WebException)
-						{
-							MessageBox.Show(
-								Strings.MessageUpdateDownloadFailed,
-								App.GetName(),
-								MessageBoxButton.OK,
-								MessageBoxImage.Error
-							);
-						}
 					}
-
+					catch (WebException)
+					{
+						m_Parent.DisplayErrorMessage(Strings.MessageUpdateDownloadFailed);
+					}
 				}
 
 			}
